Sprite shadows (#148)
* Sprite shadows

Causes crashes.

Should also try to shrink the shadow horizontally as the entity ascends above the ground.

* Shrink shadow horizontally too

* Fix crashes

With shadows enabled, `R_NewVisSprite()` is called twice pretty much back-to-back, making it possible for the first call to return a pointer which would become dangling upon the second call due to resizing of `vissprites[]`.

* Initialize tranmap only if feature is enabled

* Update docs